+++ demos/descriptions/arrayDataIndex.tsx
@@ -1,5 +1,5 @@
 ﻿import type { ProDescriptionsActionType } from '@ant-design/pro-components';
-import { ProDescriptions, ProDescriptionsItem } from '@ant-design/pro-components';
+import { ProDescriptions } from '@ant-design/pro-components';
 import { Button } from 'antd';
 import { useRef } from 'react';
 
@@ -22,46 +22,50 @@ const Demo = () => {
         });
       }}
       editable={{
-        onSave: async (keypath, newInfo, oriInfo) => {
-
+        onSave: async (_keypath, _newInfo, _oriInfo) => {
           return true;
         },
       }}
-    >
-      <ProDescriptionsItem
-        formItemProps={{
-          rules: [
-            {
-              required: true,
-              message: '此项为必填项',
-            },
+      columns={[
+        {
+          formItemProps: {
+            rules: [
+              {
+                required: true,
+                message: '此项为必填项',
+              },
+            ],
+          },
+          dataIndex: ['info', 'id'],
+        },
+        {
+          dataIndex: ['info', 'date'],
+          label: '日期',
+          valueType: 'date',
+        },
+        {
+          label: 'money',
+          dataIndex: ['info', 'money'],
+          valueType: 'money',
+        },
+        {
+          label: '文本',
+          valueType: 'option',
+          render: () => [
+            <Button
+              type="primary"
+              onClick={() => {
+                actionRef.current?.reload();
+              }}
+              key="reload"
+            >
+              刷新
+            </Button>,
+            <Button key="rest">重置</Button>,
           ],
-        }}
-        dataIndex={['info', 'id']}
-      />
-      <ProDescriptionsItem
-        dataIndex={['info', 'date']}
-        label="日期"
-        valueType="date"
-      />
-      <ProDescriptionsItem
-        label="money"
-        dataIndex={['info', 'money']}
-        valueType="money"
-      />
-      <ProDescriptionsItem label="文本" valueType="option">
-        <Button
-          type="primary"
-          onClick={() => {
-            actionRef.current?.reload();
-          }}
-          key="reload"
-        >
-          刷新
-        </Button>
-        <Button key="rest">重置</Button>
-      </ProDescriptionsItem>
-    </ProDescriptions>
+        },
+      ]}
+    />
   );
 };